analise de sentimento da descricao, corrigindo layout de card para listar feedbacks

+++ func/create_feedback.py
@@ -6,12 +6,13 @@ import psycopg2
 from db import insert_itens
 from transformers import pipeline
 from googletrans import Translator
+import datetime
 
-pipe = pipeline(model="cardiffnlp/twitter-roberta-base-sentiment-latest")
+pipe = pipeline(model="distilbert-base-uncased-finetuned-sst-2-english")
 
-st.title("Cadastre Feedback")
+st.title("Cadastre Feedback de um Local")
 
-fileupload = st.file_uploader("Escolha sua foto", type=['png', 'jpg'])
+fileupload = st.file_uploader("Foto do local", type=['png', 'jpg'])
 
 
 if fileupload is not None:
@@ -21,41 +22,50 @@ if fileupload is not None:
 
 
 
-city, name, col3 = st.columns(3)
+city, name, data = st.columns(3)
 with city:
     cidade = st.selectbox("Cidade do Local", get_cities())
 
 with name:
-    nome = st.text_input("Seu nome", "")
+    nome = st.text_input("Nome do local", "", placeholder="Praça da Sé")
 
-    street = st.text_input("Rua", "")
+with data:
+    today_data = st.date_input("Data que a foto foi tirada", datetime.date(2023, 8, 18))
 
-    opinion = st.text_area("Descreva sobre esse local", "")
-    cont = len(opinion)
+street = st.text_input("Rua", "", placeholder="Rua, Bairro")
+
+opinion = st.text_area("Descreva sobre esse local", "", placeholder="Este local é...")
+cont = len(opinion)
 
 st.write("Quantidade de caracteries: ", cont)
 
 
 
-submit = st.button("enviar")
+submit = st.button("Enviar Feedback")
 
 #apaga foto do arquivo após ser enviada
 if(submit):
-    caminho = os.path.abspath(fileupload.name)
-    with open(caminho, 'rb') as img_file:
-        imagem_data = img_file.read()
+    if fileupload:
+        caminho = os.path.abspath(fileupload.name)
+        with open(caminho, 'rb') as img_file:
+            imagem_data = img_file.read()
 
-    hash_img = psycopg2.Binary(imagem_data)
-    
-    if hash_img and city and street and opinion:
-        translator = Translator()
-        translated_opinion = translator.translate(opinion, dest="en")
+        hash_img = psycopg2.Binary(imagem_data)
+        
+        if hash_img and city and street and opinion and today_data:
+            translator = Translator()
+            translated_opinion = translator.translate(opinion, dest="en")
+            sentiment = pipe(translated_opinion.text)[0]['label']
+            res = insert_itens(nome, hash_img, cidade, street, opinion, sentiment.title(), str(today_data))
 
-        sentiment = pipe(translated_opinion.text)
-
-        print(translated_opinion.text, sentiment)
-
-        #res = insert_itens(nome, hash_img, cidade, street, opinion, sentiment)
-
-    os.remove(caminho)
+            if res:
+                st.success('Feedback enviado com sucesso!', icon="✅")
+                st.balloons()
+            else:
+                st.error('Não foi possivel enviar o feedback!', icon="🚨")
+        else:
+            st.error('Há campos vazios!', icon="🚨")
+        os.remove(caminho)
+    else:
+        st.warning('Adicione uma imagem!', icon="⚠️")
